repaired bugs and add unit tests

+++ src/main/java/com/example/demo/handlers/DeleteHandler.java
@@ -5,16 +5,22 @@ import com.example.demo.enums.Result;
 import com.example.demo.models.Request;
 import com.example.demo.models.Response;
 
-public class DeleteHandler implements IHandler{
+public class DeleteHandler implements IHandler {
+
     @Override
     public Response execute(Request request, IDatabase db) {
         String userName = request.getUser();
         String taskName = request.getArg();
 
-        if(db.deleteTask(userName, taskName)){
+        if (!db.getTasks(userName).contains(taskName)) {
+            LOG.debug("Task " + taskName + " missing from the user " + userName);
+            return new Response(Result.ACCESS_DENIED);
+        }
+
+        if (db.deleteTask(userName, taskName)) {
             LOG.debug("Deleted task " + taskName);
             return new Response(Result.DELETED);
-        }else{
+        } else {
             LOG.debug("Failed delete task " + taskName);
             return new Response(Result.ERROR);
         }
